# person_dashboard-main/Workbench/shared/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Personal AI Workbench Local API
  version: 1.5.0
  description: >-
    Local-loopback MVP contract. Foundation session/security and the durable
    project, milestone, task, manual discussion-to-decision, explicit text/link
    capture, daily focus/review, controlled Markdown source intake, and authorized
    context full-text search slices are implemented; other product domains remain out of scope.
servers:
  - url: http://127.0.0.1:8787
paths:
...
  /api/v1/spaces:
    get:
      operationId: listSpaces
      security: [{ localSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/spaces/{spaceId}:
    get:
      operationId: getSpace
      security: [{ localSession: [] }]
      parameters:
        - in: path
          name: spaceId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/projects:
    get:
      operationId: listProjects
      security: [{ localSession: [] }]
      parameters:
        - in: query
          name: space_id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/ProjectStatus" }
        - in: query
          name: template_type
          schema: { $ref: "#/components/schemas/ProjectTemplate" }
        - in: query
          name: cursor
          schema: { type: string, maxLength: 512 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: createProject
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectCreate" }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}:
    get:
      operationId: getProject
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateProject
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectPatch" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteProject
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/transitions:
    post:
      operationId: transitionProject
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectTransition" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/restore:
    post:
      operationId: restoreProject
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "410": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/milestones:
    get:
      operationId: listMilestones
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: createMilestone
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MilestoneCreate" }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/milestones/{milestoneId}:
    patch:
      operationId: updateMilestone
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/MilestoneId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MilestonePatch" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteMilestone
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/MilestoneId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/tasks:
    get:
      operationId: listTasks
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: createTask
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskCreate" }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/tasks/{taskId}:
    get:
      operationId: getTask
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateTask
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskPatch" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteTask
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/tasks/{taskId}/transitions:
    post:
      operationId: transitionTask
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskTransition" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/discussions:
    get:
      operationId: listDiscussions
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: createDiscussion
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DiscussionCreate" }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/discussions/{discussionId}/entries:
    get:
      operationId: listDiscussionEntries
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/DiscussionId"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: createDiscussionEntry
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/DiscussionId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DiscussionEntryCreate" }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/decisions:
    get:
      operationId: listDecisions
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/discussions/{discussionId}/conversions:
    post:
      operationId: convertDiscussionToDecisionAndTask
      description: >-
        Explicit owner-confirmed atomic command. It resolves one open discussion,
        accepts one decision, creates one task, and records their link. It is not an AI approval.
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/DiscussionId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DiscussionConversion" }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/captures:
    get:
      operationId: listCaptures
      security: [{ localSession: [] }]
      parameters:
        - in: query
          name: space_id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: project_id
          schema: { type: string, format: uuid }
        - in: query
          name: status
          schema: { type: string, enum: [inbox, processed, archived], default: inbox }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: createCapture
      description: >-
        Stores explicit text or http/https link metadata in the local inbox.
        The endpoint never fetches the URL and never promotes the capture to long-term knowledge.
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CaptureCreate" }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/captures/{captureId}/transitions:
    post:
      operationId: transitionCapture
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/CaptureId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CaptureTransition" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/daily/{date}:
    get:
      operationId: getDailySnapshot
      description: >-
        Returns project task truth, at most three saved focus references, and the
        explicit user-authored review for one local date. It does not copy tasks or generate AI text.
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/DailyDate"
        - in: query
          name: space_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/daily-plans/{date}:
    put:
      operationId: saveDailyPlan
      description: Saves up to three references to project tasks; task records remain owned by the project workspace.
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/DailyDate"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/OptionalIfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DailyPlanSave" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/daily-reviews/{date}:
    put:
      operationId: saveDailyReview
      description: Stores only the user's explicit review fields; no AI generation or silent rewriting occurs.
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/DailyDate"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/OptionalIfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DailyReviewSave" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    localSession:
      type: apiKey
      in: cookie
      name: workbench_session
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string, minLength: 32 }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 16, maxLength: 200, pattern: "^[A-Za-z0-9._~-]+$" }
    IfMatch:
      in: header
      name: If-Match
      required: true
      schema: { type: string, pattern: '^"v[1-9][0-9]*"$' }
...
    ProjectId:
      in: path
      name: projectId
      required: true
      schema: { type: string, format: uuid }
    MilestoneId:
      in: path
      name: milestoneId
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Success:
      description: Successful envelope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessEnvelope" }
    Error:
      description: Safe error envelope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    ProjectStatus:
      type: string
      enum: [draft, active, paused, completed, archived]
    ProjectTemplate:
      type: string
      enum: [general, research, ai_exploration, frontier_tracking, learning]
...
    ProjectTransition:
      type: object
      additionalProperties: false
      required: [action]
      properties:
        action: { type: string, enum: [activate, pause, complete, archive, reopen] }
...